test(mcp): split oversized mcp suites
