+++ tools/dev/e2e.ts
@@ -41,7 +41,7 @@ const E2E_WORKER_SOURCE_ENTRIES = [
   },
 ] as const;
 const E2E_WORKER_CONTRACT_FILES = [
-  path.join("cloudflare", "_worker.js"),
+  "_worker.js",
   path.join("cloudflare-tmp", "manifest.js"),
   path.join("output", "server", "index.js"),
 ] as const;
@@ -62,6 +62,7 @@ const E2E_WORKER_VAR_KEYS = [
   "METRICS_BEARER_TOKEN",
   "UPLOAD_TOTAL_QUOTA_MB",
 ] as const;
+const E2E_REQUIRED_WORKER_VAR_KEYS = ["AUTH_SECRET"] as const;
 
 function parseLocalPlaywrightBaseUrl(value: string) {
   const url = new URL(value);
@@ -244,6 +245,22 @@ function copyE2EWorkerArtifact(root: string) {
       fs.copyFileSync(sourcePath, targetPath);
     }
   }
+
+  writeE2EWorkerEntrypoint(root);
+}
+
+function writeE2EWorkerEntrypoint(root: string) {
+  const sourcePath = e2eWorkerArtifactPath(
+    root,
+    path.join("cloudflare", "_worker.js"),
+  );
+  const targetPath = e2eWorkerArtifactPath(root, "_worker.js");
+  const source = fs.readFileSync(sourcePath, "utf8");
+  const rewritten = source
+    .replaceAll("./../output/", "./output/")
+    .replaceAll("./../cloudflare-tmp/", "./cloudflare-tmp/");
+
+  fs.writeFileSync(targetPath, rewritten);
 }
 
 export function validatePlaywrightWorkerRuntime(
@@ -281,7 +298,35 @@ function pickE2EWorkerVars(env: Record<string, string>) {
   );
 }
 
-function writePlaywrightWranglerConfig(
+function assertE2EWorkerRuntimeEnv(
+  env: Record<string, string>,
+  config: {
+    hyperdrive?: Array<{
+      binding?: string;
+    }>;
+  },
+) {
+  const missing: string[] = E2E_REQUIRED_WORKER_VAR_KEYS.filter(
+    (key) => !env[key],
+  );
+  const needsHyperdrive = config.hyperdrive?.some(
+    (binding) => binding.binding === "HYPERDRIVE",
+  );
+  if (
+    needsHyperdrive &&
+    !env.CLOUDFLARE_HYPERDRIVE_LOCAL_CONNECTION_STRING_HYPERDRIVE
+  ) {
+    missing.push("CLOUDFLARE_HYPERDRIVE_LOCAL_CONNECTION_STRING_HYPERDRIVE");
+  }
+
+  if (missing.length > 0) {
+    throw new Error(
+      `Missing E2E Worker environment variables: ${missing.join(", ")}. Load .env or pass DATABASE_URL and AUTH_SECRET before starting Playwright E2E.`,
+    );
+  }
+}
+
+export function writePlaywrightWranglerConfig(
   root: string,
   baseUrl: string,
   env: Record<string, string>,
@@ -294,11 +339,18 @@ function writePlaywrightWranglerConfig(
     main?: string;
     routes?: unknown;
     vars?: Record<string, string>;
+    hyperdrive?: Array<{
+      binding?: string;
+      id?: string;
+      localConnectionString?: string;
+    }>;
   }>(sourceConfigPath);
+
+  assertE2EWorkerRuntimeEnv(env, config);
 
   delete config.routes;
   const artifactRoot = e2eWorkerArtifactPath(root);
-  config.main = path.resolve(artifactRoot, "cloudflare", "_worker.js");
+  config.main = path.resolve(artifactRoot, "_worker.js");
   config.assets = {
     ...config.assets,
     directory: path.resolve(artifactRoot, "cloudflare"),
@@ -315,6 +367,15 @@ function writePlaywrightWranglerConfig(
         key,
         path.resolve(root, value),
       ]),
+    );
+  }
+  const hyperdriveLocalConnection =
+    env.CLOUDFLARE_HYPERDRIVE_LOCAL_CONNECTION_STRING_HYPERDRIVE;
+  if (hyperdriveLocalConnection && config.hyperdrive) {
+    config.hyperdrive = config.hyperdrive.map((binding) =>
+      binding.binding === "HYPERDRIVE"
+        ? { ...binding, localConnectionString: hyperdriveLocalConnection }
+        : binding,
     );
   }
