+++ vitest.integration.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
   test: {
     environment: "node",
     include: ["tests/integration/**/*.test.ts"],
+    fileParallelism: false,
     globals: true,
     testTimeout: 30_000,
     hookTimeout: 30_000,
